+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add `textDocument/rename` and `textDocument/prepareRename` support for `buf lsp serve`.
+- Fix panic in LSP for empty option paths.
 
 ## [v1.61.0] - 2025-11-25