+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `textDocument/rename` and `textDocument/prepareRename` support for `buf lsp serve`.
 - Fix panic in LSP for empty option paths.
+- Fix support for multi-arch image manifests for `buf beta registry plugin push`.
 
 ## [v1.61.0] - 2025-11-25