+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.62.0] - 2025-12-29
 
 - Add `swift_prefix` to managed mode.
 - Add `textDocument/rename` and `textDocument/prepareRename` support for `buf lsp serve`.
@@ -1454,7 +1454,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.61.0...HEAD
+[v1.62.0]: https://github.com/bufbuild/buf/compare/v1.61.0...v1.62.0
 [v1.61.0]: https://github.com/bufbuild/buf/compare/v1.60.0...v1.61.0
 [v1.60.0]: https://github.com/bufbuild/buf/compare/v1.59.0...v1.60.0
 [v1.59.0]: https://github.com/bufbuild/buf/compare/v1.58.0...v1.59.0